add layer functionality

+++ Slate/StrokeDepth.swift
@@ -5,12 +5,11 @@ import Foundation
 /// Newer strokes (higher `depthID`) map closer to the camera (smaller depth value).
 enum StrokeDepth {
     static let slotCount: UInt32 = 1 << 24
-    static let denominator: Float = Float(slotCount) + 1.0
+    static let denominator: Double = Double(slotCount) + 1.0
 
     static func metalDepth(for depthID: UInt32) -> Float {
         let clamped = min(depthID, slotCount - 1)
-        let numerator = Float(slotCount - clamped)
-        return numerator / denominator
+        let numerator = Double(slotCount - clamped)
+        return Float(numerator / denominator)
     }
 }
-
